+++ src/app/shared/cohort/cohort.service.spec.ts
@@ -3,8 +3,12 @@ import { TestBed } from '@angular/core/testing';
 import { CohortService } from './cohort.service';
 import { SharedModule } from '../shared.module';
 import { last } from 'rxjs/operators';
-import tenObservationBundle from '../../modules/step-2-define-cohort-page/test-fixtures/observations-10.json';
-import examplePatient from '../../modules/step-2-define-cohort-page/test-fixtures/example-patient.json';
+import tenObservationBundle
+  from '../../modules/step-2-define-cohort-page/test-fixtures/observations-10.json';
+import tenPatientBundle
+  from '../../modules/step-2-define-cohort-page/test-fixtures/patients-10.json';
+import examplePatient
+  from '../../modules/step-2-define-cohort-page/test-fixtures/example-patient.json';
 import {
   HttpTestingController,
   HttpClientTestingModule
@@ -97,9 +101,116 @@ describe('CohortService', () => {
       )
       .flush({
         entry: patientIds.map((patientId) => ({
-          resource: { ...examplePatient, id: patientId }
+          resource: {...examplePatient, id: patientId}
         }))
       });
+  });
+
+
+  it('should correctly process nested ANDed criteria if parent nodes are ORed', (done) => {
+    const criteria: Criteria = {
+      'condition': 'and',
+      'rules': [
+        {
+          'condition': 'and',
+          'rules': [
+            {
+              'field': {
+                'element': 'code text',
+                'value': '',
+                'selectedObservationCodes': {
+                  'coding': [
+                    {
+                      'code': '72166-2',
+                      'system': 'http://loinc.org'
+                    }
+                  ],
+                  'datatype': 'CodeableConcept',
+                  'items': [
+                    'Tobacco smoking status'
+                  ]
+                }
+              }
+            }
+          ],
+          'resourceType': 'Observation'
+        },
+        {
+          'condition': 'or',
+          'rules': [
+            {
+              'condition': 'and',
+              'rules': [
+                {
+                  'field': {
+                    'element': 'active',
+                    'value': 'false'
+                  }
+                }
+              ],
+              'resourceType': 'Patient'
+            },
+            {
+              'condition': 'and',
+              'rules': [
+                {
+                  'field': {
+                    'element': 'gender',
+                    'value': {
+                      'codes': [
+                        'male'
+                      ],
+                      'items': [
+                        'Male'
+                      ]
+                    }
+                  }
+                }
+              ],
+              'resourceType': 'Patient'
+            }
+          ]
+        }
+      ]
+    };
+
+    cohort.searchForPatients(criteria, 20);
+    cohort.patientStream.pipe(last()).subscribe((patients) => {
+      expect(patients.length).toEqual(10);
+      done();
+    });
+
+    mockHttp
+      .expectOne(
+        '$fhir/Patient?_total=accurate&_summary=count&_has:Observation:subject:combo-code=http%3A%2F%2Floinc.org%7C72166-2'
+      )
+      .flush({total: 10});
+
+    mockHttp
+      .expectOne('$fhir/Patient?_total=accurate&_summary=count&active=false')
+      .flush({total: 20});
+
+    mockHttp
+      .expectOne('$fhir/Patient?_total=accurate&_summary=count&gender=male')
+      .flush({total: 30});
+
+    mockHttp
+      .expectOne(
+        '$fhir/Patient?_count=20&_has:Observation:subject:combo-code=http%3A%2F%2Floinc.org%7C72166-2'
+      )
+      .flush(tenPatientBundle);
+
+    tenPatientBundle.entry.forEach(({resource}) => {
+      mockHttp
+        .expectOne(`$fhir/Patient?_id=${resource.id}&active=false`)
+        .flush({total: 0});
+    });
+
+    tenPatientBundle.entry.forEach(({resource}) => {
+      mockHttp
+        .expectOne(`$fhir/Patient?_id=${resource.id}&gender=male`)
+        .flush({total: 1, entry: [{resource}]});
+    });
   });
 
   it('should update old format criteria for observationDataType', () => {